Moved schema transform classes to the separate folder
