+++ sdks/java/io/singlestore/src/main/java/org/apache/beam/sdk/io/singlestore/schematransform/SingleStoreSchemaTransformWriteConfiguration.java
@@ -15,10 +15,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.io.singlestore;
+package org.apache.beam.sdk.io.singlestore.schematransform;
 
 import com.google.auto.value.AutoValue;
 import javax.annotation.Nullable;
+
+import org.apache.beam.sdk.io.singlestore.SingleStoreIO;
+import org.apache.beam.sdk.io.singlestore.schematransform.AutoValue_SingleStoreSchemaTransformWriteConfiguration;
 import org.apache.beam.sdk.schemas.AutoValueSchema;
 import org.apache.beam.sdk.schemas.annotations.DefaultSchema;
 import org.apache.beam.sdk.transforms.SerializableFunction;
@@ -26,30 +29,27 @@ import org.apache.beam.sdk.values.Row;
 import org.apache.beam.sdk.values.TypeDescriptor;
 
 /**
- * Configuration for parallel reading from SignleStoreDB.
+ * Configuration for writing to SignleStoreDB.
  *
- * <p>This class is meant to be used with {@link
- * SingleStoreSchemaTransformReadWithPartitionsProvider}.
+ * <p>This class is meant to be used with {@link SingleStoreSchemaTransformWriteProvider}.
  */
 @DefaultSchema(AutoValueSchema.class)
 @AutoValue
-public abstract class SingleStoreSchemaTransformReadWithPartitionsConfiguration {
+public abstract class SingleStoreSchemaTransformWriteConfiguration {
 
-  /** Instantiates a {@link SingleStoreSchemaTransformReadWithPartitionsConfiguration.Builder}. */
+  /** Instantiates a {@link SingleStoreSchemaTransformWriteConfiguration.Builder}. */
   public static Builder builder() {
-    return new AutoValue_SingleStoreSchemaTransformReadWithPartitionsConfiguration.Builder();
+    return new AutoValue_SingleStoreSchemaTransformWriteConfiguration.Builder();
   }
 
   private static final AutoValueSchema AUTO_VALUE_SCHEMA = new AutoValueSchema();
-  private static final TypeDescriptor<SingleStoreSchemaTransformReadWithPartitionsConfiguration>
-      TYPE_DESCRIPTOR =
-          TypeDescriptor.of(SingleStoreSchemaTransformReadWithPartitionsConfiguration.class);
-  private static final SerializableFunction<
-          SingleStoreSchemaTransformReadWithPartitionsConfiguration, Row>
+  private static final TypeDescriptor<SingleStoreSchemaTransformWriteConfiguration>
+      TYPE_DESCRIPTOR = TypeDescriptor.of(SingleStoreSchemaTransformWriteConfiguration.class);
+  private static final SerializableFunction<SingleStoreSchemaTransformWriteConfiguration, Row>
       ROW_SERIALIZABLE_FUNCTION = AUTO_VALUE_SCHEMA.toRowFunction(TYPE_DESCRIPTOR);
 
   /** Serializes configuration to a {@link Row}. */
-  Row toBeamRow() {
+  public Row toBeamRow() {
     return ROW_SERIALIZABLE_FUNCTION.apply(this);
   }
 
@@ -57,13 +57,10 @@ public abstract class SingleStoreSchemaTransformReadWithPartitionsConfiguration 
   public abstract SingleStoreIO.DataSourceConfiguration getDataSourceConfiguration();
 
   @Nullable
-  public abstract String getQuery();
-
-  @Nullable
   public abstract String getTable();
 
   @Nullable
-  public abstract Integer getInitialNumReaders();
+  public abstract Integer getBatchSize();
 
   @AutoValue.Builder
   public abstract static class Builder {
@@ -72,13 +69,9 @@ public abstract class SingleStoreSchemaTransformReadWithPartitionsConfiguration 
 
     public abstract Builder setTable(String value);
 
-    public abstract Builder setQuery(String value);
+    public abstract Builder setBatchSize(Integer value);
 
-    public abstract Builder setInitialNumReaders(Integer value);
-
-    /**
-     * Builds the {@link SingleStoreSchemaTransformReadWithPartitionsConfiguration} configuration.
-     */
-    public abstract SingleStoreSchemaTransformReadWithPartitionsConfiguration build();
+    /** Builds the {@link SingleStoreSchemaTransformWriteConfiguration} configuration. */
+    public abstract SingleStoreSchemaTransformWriteConfiguration build();
   }
 }
